Adjust total calcium for albumin

Most of the protein-bound calcium is bound to albumin. In hypoalbuminaemia and hypocalcaemia, the (clinically relevant ionised) calcium level will be underestimated. Therefore, total calcium levels may be adjusted/corrected for albumin. An alternative approach is measuring ionised calcium levels directly. Calcium is adjusted using the formula: \(\text{calcium}_\text{measured} + 0.025 (\text{albumin}_\text{normal} - \text{albumin}_\text{measured}))\).

adjust_calcium_for_albumin(calcium, albumin, normal_albumin = 40)

Arguments

calcium
Measured total calcium level (mmol/l).

albumin
Measured albumin level (g/l).

normal_albumin
Normal value of albumin (g/l) (40 by default).

Value

Albumin-adjusted calcium level (mmol/l), or NA if any parameters are NA.

Caveats

Adjustments for albumin are quite poor substitutes for measuring ionised calcium.

References

Bushinsky, D. A. & Monk, R. D. Calcium. The Lancet 352, 306–311 (1998)..

See also