Estimate creatinine clearance and glomerular filtration rate using the CKD-EPI formula

The CKD-EPI formula is calculated in two stages. First, \(Z = \frac{\text{creatinine}}{\kappa}\), where \(\kappa\) is 61.9 for females and 79.6 for males. Then, if \(Z < 1\), use \(141 \times Z^(\alpha) \times 1^{-1.209} \times 0.993^{\text{age}} \times S \times A\), where \(\alpha\) is -0.329 in females and -0.411 in males, \(S\) is 1.018 in females and 1 in males, and \(A\) is 1.159 for African-Americans and 1 for non-African-Americans. However, if \(Z > 1\), use \(141 \times 1^{\alpha} \times Z^{-1.209} \times 0.993^{\text{age}} \times S \times A\), where \(\alpha\) is -0.329 in females and -0.411 in males, \(S\) is 1.018 in females and 1 in males, and \(A\) is 1.159 for African-Americans and 1 for non-African-Americans.

estimate_gfr_ckdepi(creatinine, age, is_female, is_african_american)

Arguments

creatinine
Creatinine (\(\text{μmol} \text{l}^{-1}\)).

age
Age (years).

is_female
TRUE if patient is female, FALSE if patient is male.

is_african_american
TRUE if patient is African-American, FALSE if patient is not.

Value

eGFR (\(\text{ml} \text{min}^{-1} 1.73 \text{m}^{-2}\)).

Caveats

This is the formula for when cystatine C values are unavailable. The factor 'African-American' was published as black/non-black. I do not know if this is an acceptable/usable generalisation for all people with a black skin tone (or for people of other non-Caucasian, non-African-American descent for that matter).

References

Levey, A. S. et al. A New Equation to Estimate Glomerular Filtration Rate. Ann Intern Med 150, 604 (2009).

See also