Score community-acquired pneumonia severity using CURB-65

The CURB-65 community-acquired pneumonia severity score is used to estimate mortality.

score_curb65(confusion, urea, respiratory_rate, sbp, dbp, age)

Arguments

confusion
Mental confusion (TRUE or FALSE).

urea
Blood urea nitrogen (mmol/l).

respiratory_rate
Respiratory rate (/min).

sbp
Systolic blood pressure (mmHg).

dbp
Diastolic blood pressure (mmHg).

age
Age (years).

Value

CURB-65 score (points).

Caveats

None at this time.

References

Lim, W. S. Defining community acquired pneumonia severity on presentation to hospital: an international derivation and validation study. Thorax 58, 377–382 (2003)..